+++ component/utils/loadAssets.js
@@ -3,6 +3,26 @@ import React from 'react';
 
 export const bookOn = require('../../assets/icon/book_on.png');
 export const bookOff = require('../../assets/icon/book_off.png');
+export const priceTag = require('../../assets/icon/price-tag.png');
+export const priceTagB = require('../../assets/icon/price-tag_b.png');
 export const introImage1 = {uri: "https://cdn.crowdpic.net/detail-thumb/thumb_d_F78FC0AA8923C441588C382B19DF0BF8.jpg"};
 export const introImage2 = {uri: "https://previews.123rf.com/images/romeolu/romeolu1601/romeolu160100122/50594417-%EB%88%88-%EB%B0%B0%EA%B2%BD.jpg"};
 export const introImage3 = {uri: "https://previews.123rf.com/images/kittikornphongok/kittikornphongok1505/kittikornphongok150501184/40020410-%EB%8B%A4%EC%B1%84%EB%A1%9C%EC%9A%B4-%EC%88%98%EC%B1%84%ED%99%94%EC%9E%85%EB%8B%88%EB%8B%A4-%EA%B7%B8%EB%9F%B0-%EC%A7%80-%EC%A7%88%EA%B0%90-%EB%B0%B0%EA%B2%BD%EC%9E%85%EB%8B%88%EB%8B%A4-%EB%B6%80%EB%93%9C%EB%9F%AC%EC%9A%B4-%EB%B0%B0%EA%B2%BD%EC%9E%85%EB%8B%88%EB%8B%A4-.jpg"};
+
+export const logo = require('../../assets/img/icon.png');
+export const subOn = require('../../assets/icon/subOn.png');
+export const subOff = require('../../assets/icon/subOff.png');
+export const diaryImgList = [
+  require('../../assets/icon/diary_1.png'),
+  require('../../assets/icon/diary_2.png'),
+  require('../../assets/icon/diary_3.png'),
+  require('../../assets/icon/diary_4.png'),
+  require('../../assets/icon/diary_5.png'),
+  require('../../assets/icon/diary_6.png'),
+  require('../../assets/icon/diary_7.png'),
+  require('../../assets/icon/diary_8.png'),
+  require('../../assets/icon/diary_9.png'),
+  require('../../assets/icon/diary_10.png'),
+];
+export const downArrow = require('../../assets/icon/down_arrow.png');
+export const upArrow = require('../../assets/icon/up_arrow.png');
