+++ package.json
@@ -49,7 +49,8 @@
     "react-native-tab-view": "2.15.0",
     "react-native-uuid": "1.4.9",
     "react-native-web": "~0.11.7",
-    "redux": "4.0.5"
+    "redux": "4.0.5",
+    "expo-random": "~8.2.1"
   },
   "devDependencies": {
     "@babel/core": "^7.8.6",
